+++ react-app/src/components/EmptyPanel.css
