+++ react-app/src/components/EmptyPanel.js
